inline the app engine server editor
