+++ google-cloud-tools-plugin/ultimate/src/com/google/cloud/tools/intellij/appengine/server/integration/AppEngineServerHelper.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.tools.intellij.appengine.server.integration;
 
+import com.google.cloud.tools.intellij.appengine.sdk.CloudSdkPanel;
 import com.google.cloud.tools.intellij.util.GctBundle;
 
 import com.intellij.javaee.appServerIntegrations.ApplicationServerHelper;
@@ -24,8 +25,13 @@ import com.intellij.javaee.appServerIntegrations.ApplicationServerPersistentData
 import com.intellij.javaee.appServerIntegrations.ApplicationServerPersistentDataEditor;
 import com.intellij.javaee.appServerIntegrations.CantFindApplicationServerJarsException;
 import com.intellij.javaee.oss.server.JavaeePersistentData;
+import com.intellij.openapi.options.ConfigurationException;
+
+import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
+
+import javax.swing.JComponent;
 
 /**
  * @author nik
@@ -44,6 +50,28 @@ public class AppEngineServerHelper implements ApplicationServerHelper {
   }
 
   public ApplicationServerPersistentDataEditor createConfigurable() {
-    return new AppEngineServerEditor();
+    final CloudSdkPanel cloudSdkPanel = new CloudSdkPanel();
+
+    return new ApplicationServerPersistentDataEditor<ApplicationServerPersistentData>() {
+      @Override
+      protected void resetEditorFrom(ApplicationServerPersistentData s) {
+        cloudSdkPanel.reset();
+      }
+
+      @Override
+      protected void applyEditorTo(ApplicationServerPersistentData data) {
+        try {
+          cloudSdkPanel.apply();
+        } catch(ConfigurationException ce) {
+          // do nothing
+        }
+      }
+
+      @NotNull
+      @Override
+      protected JComponent createEditor() {
+        return cloudSdkPanel.getComponent();
+      }
+    };
   }
 }
